# docs/swagger.yaml
definitions:
...
  errors.ErrorResponse:
    properties:
      code:
        type: string
      details:
        additionalProperties: {}
        type: object
      message:
        type: string
    type: object
...
  models.FeatureTypeEnum:
    enum:
    - static
    - metered
    type: string
    x-enum-varnames:
    - FeatureTypeStatic
    - FeatureTypeMetered
...
  models.HttpResponse-models_PlanFeature:
    properties:
      data:
        $ref: '#/definitions/models.PlanFeature'
      message:
        type: string
      status:
        type: integer
    type: object
...
  models.PlanFeature:
    properties:
      config: {}
      created_at:
        type: string
      created_by:
        type: string
      feature_id:
        type: string
      feature_name:
        type: string
      feature_slug:
        type: string
      feature_type:
        $ref: '#/definitions/models.FeatureTypeEnum'
      id:
        type: string
      plan_id:
        type: string
      updated_at:
        type: string
      updated_by:
        type: string
    type: object
...
  planfeatures.updatePlanFeatureRequest:
    properties:
      config:
        additionalProperties: true
        type: object
      updated_by:
        type: string
    required:
    - updated_by
    type: object
...
info:
  contact: {}
paths:
...
  /v1/plans/{planID}/features/{featureID}:
    delete:
      consumes:
      - application/json
      description: Removes a feature from a plan
      parameters:
      - description: Tenant slug
        in: header
        name: X-Tenant
        required: true
        type: string
      - description: Plan ID
        format: uuid
        in: path
        name: planID
        required: true
        type: string
      - description: Feature ID
        format: uuid
        in: path
        name: featureID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: Plan feature deleted successfully
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "404":
          description: Plan feature not found
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      summary: Delete plan feature
      tags:
      - Plan Features
    put:
      consumes:
      - application/json
      description: Updates configuration of a feature associated with a plan
      parameters:
      - description: Tenant slug
        in: header
        name: X-Tenant
        required: true
        type: string
      - description: Plan ID
        format: uuid
        in: path
        name: planID
        required: true
        type: string
      - description: Feature ID
        format: uuid
        in: path
        name: featureID
        required: true
        type: string
      - description: Updated plan feature details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/planfeatures.updatePlanFeatureRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Plan feature updated successfully
          schema:
            $ref: '#/definitions/models.HttpResponse-models_PlanFeature'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "404":
          description: Plan feature not found
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/errors.ErrorResponse'
      tags:
      - Plan Features
...
swagger: "2.0"